server/requirements.txt
h11
pip
attrs
distro
wheel
typing_extensions
setuptools
anyio
sniffio
pydantic
httpx
pydantic_core
requests
idna
multidict
exceptiongroup
black
platformdirs
tomli
packaging
pathspec
click
httpcore
certifi
aiohttp
Django
asgiref
sqlparse
openai
cffi
pygit2
yarl
aiosignal
frozenlist
urllib3
djangorestframework
pycparser
python-dotenv
groq
gunicorn
whitenoise
pytz
PyYAML
inflection
uritemplate
drf-yasg
django-cors-headers
psycopg2-binary